+++ CHANGELOG.md
@@ -4,10 +4,9 @@ All notable changes to this project will be documented in this file.
 ## [4.1.0-SNAPSHOT]
 
 ### Added
-- Handle ResourceUpdateMessage: Update the local copy of resource upon receiving a ResourceUpdateMessage. If immediate update of resource fails, schedule in a queue.
+- Handle ResourceUpdateMessage: Update the local copy of resource upon receiving a ResourceUpdateMessage.
 - Add attribute for endpoint documentation reference to `ResourceMetadata`.
-- Store `ownerURI` and `originalUUID` in `RequestedResource`.
-- Store `contract` URI in `ResourceRepresentation`.
+- Store `ownerURI`, `contractID`, `artifactID`, and `originalUUID` in `RequestedResource`.
 - Add support for query params and additional headers when requesting artifacts.
 - Add input validation for query params and headers.
 - Add usage control framework checking to the classes `PolicyEnforcement` and `PolicyHandler`.